// backend/src/db/schemas/user-credentials.ts
// Code generated by automation script, DO NOT EDIT.
// Automated by pulling database and generating zod schema
// To update. Just run npm run generate:schema
// Written by akhilmhdh.

import { z } from "zod";

import { TImmutableDBKeys } from "./models";

export const UserCredentialsSchema = z.object({
  id: z.string().uuid(),
  type: z.string(),
  name: z.string(),
  encryptedData: z.string(),
  userId: z.string().uuid(),
  organizationId: z.string().uuid(),
  createdAt: z.date(),
  updatedAt: z.date()
});

export type TUserCredentials = z.infer<typeof UserCredentialsSchema>;
export type TUserCredentialsInsert = Omit<z.input<typeof UserCredentialsSchema>, TImmutableDBKeys>;
export type TUserCredentialsUpdate = Partial<Omit<z.input<typeof UserCredentialsSchema>, TImmutableDBKeys>>;